requirements.txt
asgiref==3.9.1
Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.15.0

# AI Services
langchain>=0.3,<0.4
langchain-core>=0.3,<0.4
langchain-openai>=0.3,<0.4
langchain-anthropic>=0.3,<0.4

# Client Libraries (Required by the partner packages)
openai>=1.10.0,<2.0.0
anthropic>=0.30.0,<1.0.0

# Environment Management
python-decouple==3.8
python-dotenv==1.0.0

# API Documentation
drf-spectacular==0.27.0

# Background Tasks
celery==5.3.0
redis==5.0.0

# Security & Encryption
cryptography==42.0.8
django-cryptography==1.1

# Monitoring & Health Checks
django-health-check==3.17.0
sentry-sdk==1.38.0

# Rate Limiting & Circuit Breaker
django-ratelimit==4.1.0
pybreaker==0.8.0

# Content Filtering & Validation
bleach==6.1.0
profanity-check==1.0.3

# Performance & Caching
django-redis==5.4.0
gunicorn==21.2.0


factory-boy==3.3.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0


requests==2.31.0
pillow==10.1.0